Apply suggested changes from deliverable3
